requirements.txt

dlt[duckdb]          # For data loading, includes duckdb driver
dbt-core             # dbt core functionality
dbt-duckdb           # dbt adapter for DuckDB
dagster              # Orchestrator core
dagster-dlt          # Dagster integration for dlt
dagster-dbt          # Dagster integration for dbt
dagster-webserver    # For the Dagit UI

# Dashboard
streamlit            # For the dashboard application
pandas               # Often used by Streamlit/DuckDB for dataframes
plotly
# pyarrow            # Often good for performance with pandas/duckdb, keep if needed

# API interaction & parsing
requests             # For making HTTP requests to APIs
ijson                # For efficiently parsing large/streaming JSON (snapshot)
python-dotenv        # For managing secrets (.env files)

# Platform-specific (optional, needed by dlt on Windows)
win-precise-time ; sys_platform == 'win32'